requirements.txt
fastapi
uvicorn
llama-cpp-python
faiss-cpu
sentence-transformers
PyMuPDF
langchain
python-multipart
pytest